trendbriefai-engine/requirements.txt
# TrendBrief AI Service Dependencies
feedparser==6.0.11
newspaper3k==0.2.8
beautifulsoup4==4.12.3
transformers==4.44.2
sentence-transformers==3.0.1
numpy==1.26.4
motor==3.6.0
pymongo==4.9.2
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
ollama==0.3.3
redis==5.1.1
python-dotenv==1.0.1
httpx==0.27.2
faiss-cpu==1.8.0

# Dev dependencies
hypothesis==6.112.1
